+++ src/main/java/org/example/repository/Repository.java
@@ -5,8 +5,8 @@ import java.util.List;
 public interface Repository<T, Integer> {
     void create(T object);
     void deleteAll();
-    void deleteById(int id);
-    void update(int id, T newObject);
-    T getById(int id);
+    void deleteById(long id);
+    void update(long id, T newObject);
+    T getById(long id);
     List<T> getAll();
 }
